Change deep linking implementation
